+++ public/class-tmsm-frontend-optimizations-public.php
@@ -969,4 +969,20 @@ class Tmsm_Frontend_Optimizations_Public {
 		return str_replace( 'feature=oembed', 'feature=oembed&modestbranding=1&showinfo=0&rel=0', $html );
 	}
 
+	/**
+	 * Filters the HTML returned by the oEmbed provider.
+	 *
+	 * @param string|false $data The returned oEmbed HTML (false if unsafe).
+	 * @param string       $url  URL of the content to be embedded.
+	 * @param array        $args Optional arguments, usually passed from a shortcode.
+	 *
+	 * @return string
+	 */
+	function oembed_result_nosnippet( string $data, string $url, array $args ) {
+
+		$data = '<div data-nosnippet="true">' . $data . '</div>';
+
+		return $data;
+	}
+
 }