+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * **New** - New Feature to replace WC Variations Radio Buttons plugin; All template overriding woocommerce/single-product/add-to-cart/variable.php should be removed
 * **New** - WooCommerce: Product meta includes flat rate if available (testing: display none for non admins)
 * **New** - oEmbed: add data-nosnippet on all oembed contents
+* **Fix** - Allow Polylang translation on shipping methods and local pickup in dynamic product meta
 
 ### 1.2.6: December 2nd, 2020
 * **Tweak** - Update FR translations